% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(dat, family, model = "clm", return_what = TRUE)
}
\arguments{
\item{family}{family should be one of binomial, continuous}

\item{model}{model should be one of sampler or clm (default is clm)}

\item{return_what}{return the fitted proximal control variable or the full model}
}
\description{
Main estimation funciton
 @param dat data set
}
