% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output_stan.R
\name{model_output_stan}
\alias{model_output_stan}
\title{model_output_stan}
\usage{
model_output_stan(N, C, O, T, Y)
}
\arguments{
\item{family}{family should be one of logistic, continuous (can be inferred)}

\item{model}{model should be one of sampler or clm (default is clm)}
}
\description{
Main estimation funciton
 @param dat data set
}
